Auth: Extracted api methods and removed unused imports
