+++ frontend/src/redux/Auth/AuthSlice.tsx
@@ -1,5 +1,3 @@
-import { createApi } from "@reduxjs/toolkit/query/react";
-import axiosBaseQuery from "../../client/customBaseQuery";
 import {User} from "../../types/Todo";
 import {createSlice, PayloadAction} from "@reduxjs/toolkit";
 
@@ -22,24 +20,5 @@ export const authSlice = createSlice({
     }
 })
 
-
-export const authApi = createApi({
-    reducerPath: 'authApi',
-    baseQuery: axiosBaseQuery(),
-    endpoints: (builder) => ({
-        login: builder.mutation<{token: string | undefined}, string>({
-            query: (token) => ({
-                url: 'auth/google',
-                method: 'POST',
-                data: JSON.stringify({token}),
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            })
-        })
-    })
-});
-
 export const { setUser, removeUser } = authSlice.actions;
-export const { useLoginMutation } = authApi;
 export default authSlice.reducer;
